backend/requirements/requirements.txt
Django==3.2.10
django-graphql-jwt==0.3.1
django-cors-headers==3.5.0
djangorestframework==3.12.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
django-guardian==2.4.0
psycopg2>=2.8
pyJWT==1.7.1
PyYAML==5.4.1
django-environ==0.8.1